% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{a_b_one}
\alias{a_b_one}
\title{Compute a right-multiplication with the ones matrix}
\usage{
a_b_one(A, B)
}
\arguments{
\item{A}{A square n by n matrix.}

\item{B}{A square n by n matrix.}
}
\value{
A * (B %*% one_mat)
}
\description{
Compute A * (B %*% one_mat) where A, B, ones_mat are square matrices of the same size,
and one_mat is a ones matrix. The product * is an entry-wise (Hadamard) product,
while %*% represents matrix multiplication.
This method is more efficient than the naive approach when A or B are sparse.
}
