% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{a_one_b}
\alias{a_one_b}
\title{Compute a left-multiplication with the ones matrix}
\usage{
a_one_b(A, B)
}
\arguments{
\item{A}{A square n by n matrix.}

\item{B}{A square n by n matrix.}
}
\value{
A * (one_mat %*% B)
}
\description{
Compute A * (one_mat %*% B) where A, B, ones_mat are square matrices of the same size,
and one_mat is a ones matrix. The product * is an entry-wise (Hadamard) product,
while %*% represents matrix multiplication.
This method is more efficient than the naive approach when A or B are sparse.
}
