% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{build_G}
\alias{build_G}
\title{Build sparse adjacency matrix}
\usage{
build_G(adj_mat)
}
\arguments{
\item{adj_mat}{Original adjacency matrix.}
}
\value{
The adjacency matrix G in sparse form.
}
\description{
Build the sparse adjacency matrix G from a graph adjacency matrix.
}
