% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{build_Gp}
\alias{build_Gp}
\title{Build product adjacency matrix}
\usage{
build_Gp(adj_mat)
}
\arguments{
\item{adj_mat}{Original adjacency matrix.}
}
\value{
A product adjacency matrix Jp in sparse form.
}
\description{
Build the sparse product adjacency matrix Jp from a
graph adjacency matrix.
}
