% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{build_Gs}
\alias{build_Gs}
\title{Build single-edge adjacency matrix}
\usage{
build_Gs(adj_mat)
}
\arguments{
\item{adj_mat}{Original adjacency matrix.}
}
\value{
A single-edge adjacency matrix Gs in sparse form.
}
\description{
Build the sparse single-edge adjacency matrix Gs from a graph adjacency matrix.
}
