% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifadjacency.R
\name{build_indicator_matrices}
\alias{build_indicator_matrices}
\title{Build indicator matrices}
\usage{
build_indicator_matrices(adj_mat)
}
\arguments{
\item{adj_mat}{Adjacency matrix from which to build the indicator matrices.}
}
\value{
A list with 8 entries:
G is the original adjacency matrix;
J is the directed indicator matrix;
J0 is the missing-edge indicator matrix;
Jn is the vertex-distinct indicator matrix;
Gs is the single-edge adjacency matrix;
Gd is the double-edge adjacency matrix
Js is the single-edge indicator matrix;
Jd is the double-edge indicator matrix.
}
\description{
Build the indicator matrices and variants of
adjacency matrices from a graph adjacency matrix.
}
\keyword{adjacency}
\keyword{indicator}
\keyword{matrix}
