% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifadjacency.R
\name{drop0_killdiag}
\alias{drop0_killdiag}
\title{Set diagonal entries to zero and sparsify}
\usage{
drop0_killdiag(mat)
}
\arguments{
\item{mat}{A matrix.}
}
\value{
A sparse-form copy of mat with its diagonal entries set to zero.
}
\description{
Set the diagonal entries of a matrix to zero
and convert it to sparse matrix form.
}
\keyword{diagonal}
\keyword{matrix}
\keyword{sparse}
