% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{get_first_eigs}
\alias{get_first_eigs}
\title{Compute first few eigenvalues and eigenvectors.}
\usage{
get_first_eigs(mat, l)
}
\arguments{
\item{mat}{The symmetric matrix for which eigenvalues and eigenvectors are to be calculated.}

\item{l}{The number of eigenvalues and eigenvectors to calculate.}
}
\value{
The first \code{l} eigenvalues (by magnitude) and associated eigenvectors of \code{mat}.

A list with two entries: \code{vals} contains the a vector of the first few eigenvalues,
and \code{vects} contains a matrix of the associated eigenvectors.
}
\description{
Compute the first few eigenvalues (by magnitude) and
associated eigenvectors of a matrix.
}
\examples{
get_first_eigs(matrix(rep(1,9), nrow=3), 2)
}
\keyword{eigenvalue}
\keyword{eigenvector}
\keyword{matrix}
\keyword{spectrum}
