% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_largest_component}
\alias{get_largest_component}
\title{Get largest connected component}
\usage{
get_largest_component(adj_mat)
}
\arguments{
\item{adj_mat}{An adjacency matrix of a graph.}
}
\value{
A vector of indices corresponding to the vertices in the largest
connected component.
}
\description{
Get the indices of the vertices in the largest connected
component of a graph from its adjacency matrix.
}
\examples{
adj_mat <- matrix(c(0, 1, 0, 0, 0, 0, 0, 0, 0), nrow = 3)
get_largest_component(adj_mat)
}
