% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_motif_names}
\alias{get_motif_names}
\title{Get common motif names}
\usage{
get_motif_names()
}
\value{
A vector of names (strings) of common motifs.
}
\description{
Get the names of some common motifs as strings.
}
