% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifadjacency.R
\name{mam_M3}
\alias{mam_M3}
\title{Perform the motif adjacency matrix calculations for motif M3}
\usage{
mam_M3(adj_mat, motif_type, weight_type)
}
\arguments{
\item{adj_mat}{Adjacency matrix from which to build the motif adjacency matrix.}

\item{motif_type}{Type of motif adjacency matrix to build.}

\item{weight_type}{The weighting scheme to use. One of "unweighted", "mean" or "product".}
}
\value{
A motif adjacency matrix.
}
\description{
Perform the motif adjacency matrix calculations for motif M3
}
