% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifadjacency.R
\name{mam_M5}
\alias{mam_M5}
\title{Perform the motif adjacency matrix calculations for motif M5}
\usage{
mam_M5(adj_mat, motif_type, mam_weight_type)
}
\arguments{
\item{adj_mat}{Adjacency matrix from which to build the motif
adjacency matrix.}

\item{motif_type}{Type of motif adjacency matrix to build.}

\item{mam_weight_type}{The weighting scheme to use.
One of \code{"unweighted"}, \code{"mean"} or \code{"product"}.}
}
\value{
A motif adjacency matrix.
}
\description{
Perform the motif adjacency matrix calculations for motif M5
}
\keyword{internal}
