% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifadjacency.R
\name{mam_Md}
\alias{mam_Md}
\title{Perform the motif adjacency matrix calculations for motif Md}
\usage{
mam_Md(adj_mat, weight_type)
}
\arguments{
\item{adj_mat}{Adjacency matrix from which to build the motif adjacency matrix.}

\item{weight_type}{The weighting scheme to use. One of "unweighted", "mean" or "product".}
}
\value{
A motif adjacency matrix.
}
\description{
Perform the motif adjacency matrix calculations for motif Md
}
