% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifadjacency.R
\name{mam_Mexpa}
\alias{mam_Mexpa}
\title{Perform the motif adjacency matrix calculations for motif Mexpa}
\usage{
mam_Mexpa(adj_mat, motif_type, weight_type, method)
}
\arguments{
\item{adj_mat}{Adjacency matrix from which to build the motif
adjacency matrix.}

\item{motif_type}{Type of motif adjacency matrix to build.}

\item{weight_type}{The weighting scheme to use.
One of "unweighted", "mean" or "product".}

\item{method}{Which formulation to use. One of "dense" or "sparse".}
}
\value{
A motif adjacency matrix.
}
\description{
Perform the motif adjacency matrix calculations for motif Mexpa
}
