% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifcalcs.R
\name{run_motif_adjacency_calcs}
\alias{run_motif_adjacency_calcs}
\title{Run motif adjacency matrix calculations}
\usage{
run_motif_adjacency_calcs(
  ind_mats,
  motif_name,
  motif_type,
  weight_type,
  method
)
}
\arguments{
\item{motif_name}{The motif to use for the motif adjacency matrix.}

\item{G}{The original adjacency matrix.
Replaced by the single-edge adjacency matrix Gs for structural motifs.}

\item{J}{The directed indicator matrix.
Replaced by the single-edge indicator matrix Js for structural motifs.}

\item{Jn}{is the vertex-distinct indicator matrix.
Replaced by the missing-edge indicator matrix J0 for structural motifs.}

\item{Jd}{The double-edge indicator matrix.}

\item{Gd}{The double-edge adjacency matrix.}
}
\value{
A motif adjacency matrix.
}
\description{
Perform the matrix operations required to build a motif
adjacency matrix from adjacency and indicator matrices.
Parameter names default to building functional motif adjacency matrices.
}
\keyword{adjacency}
\keyword{calculation}
\keyword{matrix}
\keyword{motif}
\keyword{operation}
