% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral.R
\name{run_motif_embedding}
\alias{run_motif_embedding}
\title{Run motif embedding}
\usage{
run_motif_embedding(
  adj_mat,
  motif_name,
  motif_type = c("func", "struc"),
  num_eigs,
  type_lap
)
}
\arguments{
\item{adj_mat}{Adjacency matrix to be embedded.}

\item{motif_name}{Motif used for the motif adjacency matrix.}

\item{motif_type}{Type of motif adjacency matrix to use.
One of "func" or "struc".}

\item{num_eigs}{Number of eigenvalues and eigenvectors for the embedding.}

\item{type_lap}{Type of Laplacian for the embedding. One of "comb" or "rw".}
}
\value{
A list with 7 entries:
adj_mat, the original adjacency matrix;
motif_adj_mat, the motif adjacency matrix;
comps, the indices of the largest connected component of the motif adjacency matrix;
adj_mat_comps, the original adjacency matrix restricted to the largest connected
component of the motif adjacency matrix;
motif_adj_mat_comps, the motif adjacency matrix restricted to its
largest connected component;
vals, the eigenvalues associated with the Laplace embedding
of the motif adjacency matrix;
vects, the eigenvectors associated with the Laplace embedding
of the motif adjacency matrix;
}
\description{
Calculate a motif adjacency matrix for a given motif and motif type,
restrict it to its largest connected component,
and then run Laplace embedding with specified Laplacian type and
number of eigenvalues and eigenvectors.
}
\examples{
TODO
}
\keyword{adjacency}
\keyword{embedding}
\keyword{laplacian}
\keyword{matrix}
\keyword{motif}
