% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sampleBSBM}
\alias{sampleBSBM}
\title{Sample a BSBM}
\usage{
sampleBSBM(
  source_block_sizes,
  dest_block_sizes,
  bipartite_connection_matrix,
  weight_type = c("unweighted", "deterministic", "poisson"),
  bipartite_weight_matrix = NULL
)
}
\arguments{
\item{source_block_sizes}{A vector containing the size of each block of source vertices.}

\item{dest_block_sizes}{A vector containing the size of each block of dest vertices.}

\item{bipartite_connection_matrix}{A matrix containing the block-to-block
connection probabilities.}

\item{weight_type}{The type of weighting scheme.
One of "unweighted", "deterministic" or "poisson".}

\item{bipartite_weight_matrix}{A matrix containing the
sourece block-to-destination block weight parameters.
Unused for weight_type="deterministic".
Defaults to NULL.}
}
\value{
A randomly sampled (weighted) adjacency matrix of a BSBM.
}
\description{
Sample the (weighted) adjacency matrix of a (weighted) bipartite stochastic block
model (BSBM) with specified parameters.
}
\keyword{BSBM}
\keyword{bipartite}
\keyword{model}
\keyword{sample}
