% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{a_b_one}
\alias{a_b_one}
\title{Compute a right-multiplication with the ones matrix}
\usage{
a_b_one(a, b)
}
\arguments{
\item{a}{A square n by n matrix.}

\item{b}{A square n by n matrix.}
}
\value{
a * (b %*% one_mat)
}
\description{
Compute a * (b %*% one_mat) where a, b, ones_mat are
square matrices of the same size, and one_mat is a ones matrix.
The product * is an entry-wise (Hadamard) product,
while %*% represents matrix multiplication.
This method is more efficient than the naive approach
when a or b are sparse.
}
