% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{a_one_b}
\alias{a_one_b}
\title{Compute a left-multiplication with the ones matrix}
\usage{
a_one_b(a, b)
}
\arguments{
\item{a}{A square n by n matrix.}

\item{b}{A square n by n matrix.}
}
\value{
a * (one_mat %*% b)
}
\description{
Compute a * (one_mat %*% b) where a, b, ones_mat are
square matrices of the same size, and one_mat is a ones matrix.
The product * is an entry-wise (Hadamard) product,
while %*% represents matrix multiplication.
This method is more efficient than the naive approach
when a or b are sparse.
}
