% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifadjacency.R
\name{build_motif_adjacency_matrix}
\alias{build_motif_adjacency_matrix}
\title{Build a motif adjacency matrix}
\usage{
build_motif_adjacency_matrix(
  adj_mat,
  motif_name,
  motif_type = c("func", "struc"),
  weight_type = c("unweighted", "mean", "product"),
  method = c("dense", "sparse")
)
}
\arguments{
\item{adj_mat}{Adjacency matrix from which to build the motif
adjacency matrix.}

\item{motif_name}{Motif used for the motif adjacency matrix.}

\item{motif_type}{Type of motif adjacency matrix to build.}

\item{weight_type}{The weighting scheme to use.
One of "unweighted", "mean" or "product".}

\item{method}{Which formulation to use. One of "dense" or "sparse".}
}
\value{
A motif adjacency matrix.
}
\description{
Build a motif adjacency matrix from an adjacency matrix.
}
