% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifadjacency.R
\name{mam_Md}
\alias{mam_Md}
\title{Perform the motif adjacency matrix calculations for motif Md}
\usage{
mam_Md(adj_mat, mam_weight_type)
}
\arguments{
\item{adj_mat}{Adjacency matrix from which to build the motif
adjacency matrix.}

\item{mam_weight_type}{The weighting scheme to use.
One of \code{"unweighted"}, \code{"mean"} or \code{"product"}.}
}
\value{
A motif adjacency matrix.
}
\description{
Perform the motif adjacency matrix calculations for motif Md
}
\keyword{internal}
