% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifadjacency.R
\name{run_motif_adjacency_calcs}
\alias{run_motif_adjacency_calcs}
\title{Run motif adjacency matrix calculations}
\usage{
run_motif_adjacency_calcs(G, J, Jn, Jd, Gd, motif_name)
}
\arguments{
\item{G}{The original adjacency matrix.}

\item{J}{The directed indicator matrix.}

\item{Jn}{is the vertex-distinct indicator matrix.}

\item{Jd}{The double-edge indicator matrix.}

\item{Gd}{The double-edge adjacency matri.}
}
\description{
Perform the matrix operations required to build a motif
adjacency matrix from adjacency and indicator matrices.
}
