% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_dsbm}
\alias{sample_dsbm}
\title{Sample a DSBM}
\usage{
sample_dsbm(
  block_sizes,
  connection_matrix,
  weight_type = c("unweighted", "deterministic", "poisson"),
  weight_matrix = NULL
)
}
\arguments{
\item{block_sizes}{A vector containing the size of each block of vertices.}

\item{connection_matrix}{A matrix containing the block-to-block connection probabilities.}

\item{weight_type}{The type of weighting scheme.
One of "unweighted", "deterministic" or "poisson".}

\item{weight_matrix}{A matrix containing the block-to-block weight parameters.
Unused for weight_type="deterministic".
Defaults to NULL.}
}
\value{
A randomly sampled (weighted) adjacency matrix of a DSBM.
}
\description{
Sample the (weighted) adjacency matrix of a (weighted) directed stochastic block
model (DSBM) with specified parameters.
}
\keyword{DSBM}
\keyword{model}
\keyword{sample}
